TP#39273 code review fixes

+++ Code/Interapptive.Shared/AutoUpdate/AutoUpdateTelemetryCollector.cs
@@ -17,7 +17,7 @@ namespace Interapptive.Shared.AutoUpdate
         /// <summary>
         /// Start collecting telemetry
         /// </summary>
-        public static void Start(Version targetVersion)
+        public static void Start(Version targetVersion, Guid instance)
         {
             try
             {
@@ -25,6 +25,7 @@ namespace Interapptive.Shared.AutoUpdate
                 data.Started = DateTime.UtcNow;
                 data.StartVersion = GetVersionString(Assembly.GetExecutingAssembly().GetName().Version);
                 data.TargetVersion = GetVersionString(targetVersion);
+                data.Instance = instance;
 
                 File.WriteAllText(tempPath, JsonConvert.SerializeObject(data));
             }
@@ -37,22 +38,22 @@ namespace Interapptive.Shared.AutoUpdate
         /// Load telemetry from disk
         /// </summary>
         /// <returns></returns>
-        private static AutoUpdateTelemetry? Load()
+        private static AutoUpdateTelemetry? Load(Guid instance)
         {
             if(File.Exists(tempPath))
             {
                 try
                 {
                     string data = File.ReadAllText(tempPath);
-                    return JsonConvert.DeserializeObject<AutoUpdateTelemetry>(data);
+                    var telemetry = JsonConvert.DeserializeObject<AutoUpdateTelemetry>(data);
+                    if(telemetry.Instance == instance)
+                    {
+                        File.Delete(tempPath);
+                        return telemetry;
+                    }
                 }
-                catch (Exception ex)
+                catch (Exception)
                 {
-                    return null;
-                }
-                finally
-                {
-                    File.Delete(tempPath);
                 }
             }
             return null;
@@ -68,17 +69,17 @@ namespace Interapptive.Shared.AutoUpdate
         /// Collect telemetry
         /// </summary>
         /// <param name="telemetryEvent"></param>
-        public static void CollectTelemetry()
+        public static void CollectTelemetry(Guid instance)
         {
-            var data = Load();
+            var data = Load(instance);
             if (data.HasValue)
             {
-                using (var telemetryEvent = new TrackedEvent(""))
+                using (var telemetryEvent = new TrackedEvent("AutoUpdate"))
                 {
                     telemetryEvent.AddProperty("CurrentVersion", data.Value.StartVersion);
                     telemetryEvent.AddProperty("TargetVersion", data.Value.TargetVersion);
                     telemetryEvent.AddProperty("MachineName", Environment.MachineName);
-                    telemetryEvent.AddMetric("Duration", DateTime.UtcNow.Subtract(data.Value.Started).TotalMilliseconds);
+                    telemetryEvent.AddMetric("DurationMS", DateTime.UtcNow.Subtract(data.Value.Started).TotalMilliseconds);
 
                     if (Version.TryParse(data.Value.StartVersion, out Version started))
                     {
